web - change when to sort filteredCompetitions by top and non-top competitions to also keep sort order when filtering data
