+++ web/src/components/features/competitions/competitions.hooks.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from "react";
 import { AxiosResponse } from "axios";
 import { ICompetition, ICompetitionResponse } from "@/components/common/api-football-response";
-import { TOP_COMPS_IDS } from "@/components/common/types";
 import axios from "@/lib/axios";
 
 interface ICompetitionsData {
@@ -27,32 +26,8 @@ export const useCompetitions = () => {
                 if (resp.status !== 200) {
                     throw new Error();
                 }
-                const newCompetitions = resp.data.response;
-
-                const topCompetitions: ICompetition[] = [];
-                const nonTopCompetitions: ICompetition[] = [];
-
-                // group competitons by if it is a top com
-                newCompetitions.forEach((comp) => {
-                    TOP_COMPS_IDS.includes(comp.league.id) ? topCompetitions.push(comp) : nonTopCompetitions.push(comp);
-                });
-
-                // within each sub group of top and non-top comps, group by country name
-                topCompetitions.sort((a, b) => {
-                    const aCompName = `${a.country.name} ${a.league.name}`;
-                    const bCompName = `${b.country.name} ${b.league.name}`;
-                    return aCompName.localeCompare(bCompName);
-                });
-                nonTopCompetitions.sort((a, b) => {
-                    const aCompName = `${a.country.name} ${a.league.name}`;
-                    const bCompName = `${b.country.name} ${b.league.name}`;
-                    return aCompName.localeCompare(bCompName);
-                });
-
-                const competitionsDisplayList = [...topCompetitions, ...nonTopCompetitions];
-
-                setAllCompetitions(competitionsDisplayList);
-                setFilteredCompetitions(competitionsDisplayList);
+                setAllCompetitions(resp.data.response);
+                setFilteredCompetitions(resp.data.response);
                 setStatus("success");
             } catch (err) {
                 setStatus("error");
