+++ web/src/components/features/competitions/competitions.tsx
@@ -3,6 +3,8 @@ import { ErrorComponent } from "@/components/common/error/error";
 import { useCompetitions } from "./competitions.hooks";
 import { CompetitionsItemComponent } from "./competitions-item/competitions-item";
 import { CompetitionsSearchComponent } from "./competitions-search/competitions-search";
+import { ICompetition } from "@/components/common/api-football-response";
+import { TOP_COMPS_IDS } from "@/components/common/types";
 
 export const CompetitionsComponent = () => {
     const [data, isLoading, isError] = useCompetitions();
@@ -19,6 +21,29 @@ export const CompetitionsComponent = () => {
             />
         );
     }
+
+    const topCompetitions: ICompetition[] = [];
+    const nonTopCompetitions: ICompetition[] = [];
+
+    // group competitons by if it is a top com
+    data.filteredCompetitions.forEach((comp) => {
+        TOP_COMPS_IDS.includes(comp.league.id) ? topCompetitions.push(comp) : nonTopCompetitions.push(comp);
+    });
+
+    // within each sub group of top and non-top comps, group by country name
+    topCompetitions.sort((a, b) => {
+        const aCompName = `${a.country.name} ${a.league.name}`;
+        const bCompName = `${b.country.name} ${b.league.name}`;
+        return aCompName.localeCompare(bCompName);
+    });
+    nonTopCompetitions.sort((a, b) => {
+        const aCompName = `${a.country.name} ${a.league.name}`;
+        const bCompName = `${b.country.name} ${b.league.name}`;
+        return aCompName.localeCompare(bCompName);
+    });
+
+    const competitionsDisplayList = [...topCompetitions, ...nonTopCompetitions];
+
     return (
         <section className="space-y-5">
             <h3 className="text-2xl font-bold">All Competitions</h3>
@@ -29,8 +54,8 @@ export const CompetitionsComponent = () => {
             />
 
             <section className="grid grid-cols-1 lg:grid-cols-3 gap-1">
-                {data.filteredCompetitions.length > 0 ? (
-                    data.filteredCompetitions.map((competition) => (
+                {competitionsDisplayList.length > 0 ? (
+                    competitionsDisplayList.map((competition) => (
                         <CompetitionsItemComponent competition={competition} key={competition.league.id} />
                     ))
                 ) : (
